requirements.txt
phidata
yfinance
python-dotenv
yfinance
packaging
duckduckgo-search
openai
python-multipart
fastapi
uvicorn
langchain_openai
langchain_community
groq